+++ jest.config.js
@@ -1,13 +1,14 @@
 module.exports = {
-  globals: {
-    'ts-jest': {
-      tsconfig: 'tsconfig.json',
+    globals: {
+        "ts-jest": {
+            tsconfig: "tsconfig.json",
+        },
     },
-  },
-  moduleFileExtensions: ['ts', 'js'],
-  transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
-  },
-  testMatch: ['**/test/**/*.+(ts|js)', '**/__tests__/**/*.+(ts|tsx|js)'],
-  testEnvironment: 'node',
+    moduleFileExtensions: ["ts", "js"],
+    transform: {
+        "^.+\\.(ts|tsx)$": "ts-jest",
+    },
+    modulePathIgnorePatterns: [".*\\.d\\.ts"],
+    testMatch: ["**/test/**/*.+(ts)", "**/__tests__/**/*.+(ts|tsx)"],
+    testEnvironment: "node",
 }